chatgpt-article-summary/requirements.txt
discord.py==2.2.2
requests==2.22.0
openai==0.27.8
langchain==0.0.240
python-dotenv==1.0.0